// tsp-output/server/js/src/generated/models/all/syncly-service.ts
export interface ErrorResponse {
  code: string;

  message: string;

  details?: string;
}
// ...
export interface ParticipantAnswer {
  id: string;

  name: string;

  canJoinRemotely: boolean;

  comment?: string;

  availableCandidateIds: string[];

  createdAt: Temporal.Instant;

  updatedAt: Temporal.Instant;
}
// ...
export interface CreateParticipantRequest {
  name: string;

  canJoinRemotely: boolean;

  comment?: string;

  availableCandidateIds: string[];
}

export interface UpdateParticipantRequest {
  name?: string;

  canJoinRemotely?: boolean;

  comment?: string;

  availableCandidateIds?: string[];
}
// ...
export interface Participants<Context = unknown> {
  /**
   * 募集ごとの参加者回答を取得する
   */
  list(
    ctx: Context,
    pageId: string,
  ): Promise<ParticipantAnswer[] | ErrorResponse>;

  /**
   * 参加者回答を作成する
   */
  create(
    ctx: Context,
    pageId: string,
    body: CreateParticipantRequest,
  ): Promise<ParticipantAnswer | ErrorResponse>;

  /**
   * 参加者回答を取得する
   */
  read(
    ctx: Context,
    pageId: string,
    participantId: string,
  ): Promise<ParticipantAnswer | ErrorResponse>;

  /**
   * OPENの募集に対する参加者回答を更新する
   */
  update(
    ctx: Context,
    pageId: string,
    participantId: string,
    body: UpdateParticipantRequest,
  ): Promise<ParticipantAnswer | ErrorResponse>;

  /**
   * 参加者回答を削除する
   */
  delete(
    ctx: Context,
    pageId: string,
    participantId: string,
  ): Promise<void | ErrorResponse>;
}
// ...
export const ParticipantAnswer = {
  toJsonObject(input: ParticipantAnswer): any {
    return {
      id: input.id,
      name: input.name,
      canJoinRemotely: input.canJoinRemotely,
      comment: input.comment,
      availableCandidateIds: input.availableCandidateIds,
      createdAt: input.createdAt.toString(),
      updatedAt: input.updatedAt.toString(),
    };
  },
  fromJsonObject(input: any): ParticipantAnswer {
    return {
      id: input.id,
      name: input.name,
      canJoinRemotely: input.canJoinRemotely,
      comment: input.comment,
      availableCandidateIds: input.availableCandidateIds,
      createdAt: Temporal.Instant.from(input.createdAt),
      updatedAt: Temporal.Instant.from(input.updatedAt),
    };
  },
} as const;